+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.3.3+1.21.9]
+
+### Added
+
+- Added Copper Golem to HeadUtil.
+
+### Changed
+
+- Changed supported Minecraft version number.
+
 ## [1.3.3+1.21.6]
 
 ### Added